Chat Dialog stays open
